requirements.txt
Flask
sqlalchemy
flask_sqlalchemy
Flask-Login
flask_wtf
Flask-Migrate
Flask-Mail
flask-compress
flask-bootstrap
flask-nav
flask-httpauth
flask-cors
pyqt5
Jinja2
itsdangerous
pycryptodome
click
celery
pyjwt
faker
sqlalchemy-utils
marshmallow